*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@5595 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/modules/er035m_s.c
@@ -1270,7 +1270,7 @@ static bool er035m_s_comm( int type, ... )
 			   sends are completely useless, so get rid of them... */
 
 			*lptr = len;
-			for ( len = 0; len < *lptr; len++ )
+			for ( len = 0; len < ( ssize_t ) *lptr; len++ )
 				buf[ len ] &= 0x3f;
 
 			break;
